requirements.txt
appdirs==1.4.3
apturl==0.5.2
arrow==0.14.7
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.4
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2020.6.20
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
gpg===1.13.1-unknown
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
isort==5.5.4
joblib==0.17.0
jsonschema==3.2.0
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
Markdown==3.1.1
matplotlib==3.3.2
mccabe==0.6.1
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.6.0
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rope==0.17.0
scikit-learn==0.23.2
scipy==1.5.2
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
systemd-python==234
TatSu==5.5.0
texttable==1.6.2
threadpoolctl==2.1.0
toml==0.10.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.0.17
wadllib==1.3.3
websocket-client==0.53.0
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
